Remove static banner generation
