+++ versionwarning/_static/js/versionwarning.js
@@ -1,1 +1,1 @@
-!function(r){var e={};function t(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return r[n].call(o.exports,o,o.exports,t),o.l=!0,o.exports}t.m=r,t.c=e,t.d=function(r,e,n){t.o(r,e)||Object.defineProperty(r,e,{enumerable:!0,get:n})},t.r=function(r){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(r,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(r,"__esModule",{value:!0})},t.t=function(r,e){if(1&e&&(r=t(r)),8&e)return r;if(4&e&&"object"==typeof r&&r&&r.__esModule)return r;var n=Object.create(null);if(t.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:r}),2&e&&"string"!=typeof r)for(var o in r)t.d(n,o,function(e){return r[e]}.bind(null,o));return n},t.n=function(r){var e=r&&r.__esModule?function(){return r.default}:function(){return r};return t.d(e,"a",e),e},t.o=function(r,e){return Object.prototype.hasOwnProperty.call(r,e)},t.p="",t(t.s=0)}([function(r,e,t){const n=t(1);function o(r){var e=r.version;console.debug("Running version: "+e.slug);var t={project__slug:r.project.slug};$.ajax({url:r.meta.api_url+"version/",data:t,success:function(t){highest_version=function(r){var e;return $.each(r,function(r,t){n.valid(t.slug)&&(e?n.valid(t.slug)&&n.valid(e.slug)&&n.gt(t.slug,e.slug)&&(e=t):e=t)}),e}(t.results),n.valid(e.slug)&&n.valid(highest_version.slug)&&n.lt(e.slug,highest_version.slug)&&(console.debug("Highest version: "+highest_version.slug),function(r,e,t){var n=window.location.pathname.replace(r.slug,e.slug),o=$(t.banner.html);o.find("a").attr("href",n).text(e.slug);var i=$(t.banner.body_default_selector);i.length||(i=$(t.banner.body_extra_selector)),i.prepend(o)}(e,highest_version,r))},error:function(){console.error("Error loading Read the Docs active versions.")}})}$(document).ready(function(){$.ajax({url:"_static/data/versionwarning-data.json",success:function(r){document.getElementById(r.banner.id_div)?console.debug("There is already a banner added. No checking versions."):o(r)},error:function(){console.error("Error loading versionwarning-data.json")}})})},function(r,e,t){(function(t){var n;e=r.exports=H,n="object"==typeof t&&t.env&&t.env.NODE_DEBUG&&/\bsemver\b/i.test(t.env.NODE_DEBUG)?function(){var r=Array.prototype.slice.call(arguments,0);r.unshift("SEMVER"),console.log.apply(console,r)}:function(){},e.SEMVER_SPEC_VERSION="2.0.0";var o=256,i=Number.MAX_SAFE_INTEGER||9007199254740991,s=e.re=[],a=e.src=[],u=0,c=u++;a[c]="0|[1-9]\\d*";var l=u++;a[l]="[0-9]+";var p=u++;a[p]="\\d*[a-zA-Z-][a-zA-Z0-9-]*";var h=u++;a[h]="("+a[c]+")\\.("+a[c]+")\\.("+a[c]+")";var f=u++;a[f]="("+a[l]+")\\.("+a[l]+")\\.("+a[l]+")";var v=u++;a[v]="(?:"+a[c]+"|"+a[p]+")";var m=u++;a[m]="(?:"+a[l]+"|"+a[p]+")";var g=u++;a[g]="(?:-("+a[v]+"(?:\\."+a[v]+")*))";var d=u++;a[d]="(?:-?("+a[m]+"(?:\\."+a[m]+")*))";var w=u++;a[w]="[0-9A-Za-z-]+";var y=u++;a[y]="(?:\\+("+a[w]+"(?:\\."+a[w]+")*))";var b=u++,j="v?"+a[h]+a[g]+"?"+a[y]+"?";a[b]="^"+j+"$";var E="[v=\\s]*"+a[f]+a[d]+"?"+a[y]+"?",T=u++;a[T]="^"+E+"$";var $=u++;a[$]="((?:<|>)?=?)";var _=u++;a[_]=a[l]+"|x|X|\\*";var x=u++;a[x]=a[c]+"|x|X|\\*";var k=u++;a[k]="[v=\\s]*("+a[x]+")(?:\\.("+a[x]+")(?:\\.("+a[x]+")(?:"+a[g]+")?"+a[y]+"?)?)?";var S=u++;a[S]="[v=\\s]*("+a[_]+")(?:\\.("+a[_]+")(?:\\.("+a[_]+")(?:"+a[d]+")?"+a[y]+"?)?)?";var R=u++;a[R]="^"+a[$]+"\\s*"+a[k]+"$";var I=u++;a[I]="^"+a[$]+"\\s*"+a[S]+"$";var O=u++;a[O]="(?:^|[^\\d])(\\d{1,16})(?:\\.(\\d{1,16}))?(?:\\.(\\d{1,16}))?(?:$|[^\\d])";var A=u++;a[A]="(?:~>?)";var M=u++;a[M]="(\\s*)"+a[A]+"\\s+",s[M]=new RegExp(a[M],"g");var V=u++;a[V]="^"+a[A]+a[k]+"$";var N=u++;a[N]="^"+a[A]+a[S]+"$";var P=u++;a[P]="(?:\\^)";var C=u++;a[C]="(\\s*)"+a[P]+"\\s+",s[C]=new RegExp(a[C],"g");var L=u++;a[L]="^"+a[P]+a[k]+"$";var D=u++;a[D]="^"+a[P]+a[S]+"$";var q=u++;a[q]="^"+a[$]+"\\s*("+E+")$|^$";var X=u++;a[X]="^"+a[$]+"\\s*("+j+")$|^$";var z=u++;a[z]="(\\s*)"+a[$]+"\\s*("+E+"|"+a[k]+")",s[z]=new RegExp(a[z],"g");var B=u++;a[B]="^\\s*("+a[k]+")\\s+-\\s+("+a[k]+")\\s*$";var G=u++;a[G]="^\\s*("+a[S]+")\\s+-\\s+("+a[S]+")\\s*$";var Z=u++;a[Z]="(<|>)?=?\\s*\\*";for(var U=0;U<u;U++)n(U,a[U]),s[U]||(s[U]=new RegExp(a[U]));function F(r,e){if(r instanceof H)return r;if("string"!=typeof r)return null;if(r.length>o)return null;if(!(e?s[T]:s[b]).test(r))return null;try{return new H(r,e)}catch(r){return null}}function H(r,e){if(r instanceof H){if(r.loose===e)return r;r=r.version}else if("string"!=typeof r)throw new TypeError("Invalid Version: "+r);if(r.length>o)throw new TypeError("version is longer than "+o+" characters");if(!(this instanceof H))return new H(r,e);n("SemVer",r,e),this.loose=e;var t=r.trim().match(e?s[T]:s[b]);if(!t)throw new TypeError("Invalid Version: "+r);if(this.raw=r,this.major=+t[1],this.minor=+t[2],this.patch=+t[3],this.major>i||this.major<0)throw new TypeError("Invalid major version");if(this.minor>i||this.minor<0)throw new TypeError("Invalid minor version");if(this.patch>i||this.patch<0)throw new TypeError("Invalid patch version");t[4]?this.prerelease=t[4].split(".").map(function(r){if(/^[0-9]+$/.test(r)){var e=+r;if(e>=0&&e<i)return e}return r}):this.prerelease=[],this.build=t[5]?t[5].split("."):[],this.format()}e.parse=F,e.valid=function(r,e){var t=F(r,e);return t?t.version:null},e.clean=function(r,e){var t=F(r.trim().replace(/^[=v]+/,""),e);return t?t.version:null},e.SemVer=H,H.prototype.format=function(){return this.version=this.major+"."+this.minor+"."+this.patch,this.prerelease.length&&(this.version+="-"+this.prerelease.join(".")),this.version},H.prototype.toString=function(){return this.version},H.prototype.compare=function(r){return n("SemVer.compare",this.version,this.loose,r),r instanceof H||(r=new H(r,this.loose)),this.compareMain(r)||this.comparePre(r)},H.prototype.compareMain=function(r){return r instanceof H||(r=new H(r,this.loose)),K(this.major,r.major)||K(this.minor,r.minor)||K(this.patch,r.patch)},H.prototype.comparePre=function(r){if(r instanceof H||(r=new H(r,this.loose)),this.prerelease.length&&!r.prerelease.length)return-1;if(!this.prerelease.length&&r.prerelease.length)return 1;if(!this.prerelease.length&&!r.prerelease.length)return 0;var e=0;do{var t=this.prerelease[e],o=r.prerelease[e];if(n("prerelease compare",e,t,o),void 0===t&&void 0===o)return 0;if(void 0===o)return 1;if(void 0===t)return-1;if(t!==o)return K(t,o)}while(++e)},H.prototype.inc=function(r,e){switch(r){case"premajor":this.prerelease.length=0,this.patch=0,this.minor=0,this.major++,this.inc("pre",e);break;case"preminor":this.prerelease.length=0,this.patch=0,this.minor++,this.inc("pre",e);break;case"prepatch":this.prerelease.length=0,this.inc("patch",e),this.inc("pre",e);break;case"prerelease":0===this.prerelease.length&&this.inc("patch",e),this.inc("pre",e);break;case"major":0===this.minor&&0===this.patch&&0!==this.prerelease.length||this.major++,this.minor=0,this.patch=0,this.prerelease=[];break;case"minor":0===this.patch&&0!==this.prerelease.length||this.minor++,this.patch=0,this.prerelease=[];break;case"patch":0===this.prerelease.length&&this.patch++,this.prerelease=[];break;case"pre":if(0===this.prerelease.length)this.prerelease=[0];else{for(var t=this.prerelease.length;--t>=0;)"number"==typeof this.prerelease[t]&&(this.prerelease[t]++,t=-2);-1===t&&this.prerelease.push(0)}e&&(this.prerelease[0]===e?isNaN(this.prerelease[1])&&(this.prerelease=[e,0]):this.prerelease=[e,0]);break;default:throw new Error("invalid increment argument: "+r)}return this.format(),this.raw=this.version,this},e.inc=function(r,e,t,n){"string"==typeof t&&(n=t,t=void 0);try{return new H(r,t).inc(e,n).version}catch(r){return null}},e.diff=function(r,e){if(rr(r,e))return null;var t=F(r),n=F(e);if(t.prerelease.length||n.prerelease.length){for(var o in t)if(("major"===o||"minor"===o||"patch"===o)&&t[o]!==n[o])return"pre"+o;return"prerelease"}for(var o in t)if(("major"===o||"minor"===o||"patch"===o)&&t[o]!==n[o])return o},e.compareIdentifiers=K;var J=/^[0-9]+$/;function K(r,e){var t=J.test(r),n=J.test(e);return t&&n&&(r=+r,e=+e),t&&!n?-1:n&&!t?1:r<e?-1:r>e?1:0}function Q(r,e,t){return new H(r,t).compare(new H(e,t))}function W(r,e,t){return Q(r,e,t)>0}function Y(r,e,t){return Q(r,e,t)<0}function rr(r,e,t){return 0===Q(r,e,t)}function er(r,e,t){return 0!==Q(r,e,t)}function tr(r,e,t){return Q(r,e,t)>=0}function nr(r,e,t){return Q(r,e,t)<=0}function or(r,e,t,n){var o;switch(e){case"===":"object"==typeof r&&(r=r.version),"object"==typeof t&&(t=t.version),o=r===t;break;case"!==":"object"==typeof r&&(r=r.version),"object"==typeof t&&(t=t.version),o=r!==t;break;case"":case"=":case"==":o=rr(r,t,n);break;case"!=":o=er(r,t,n);break;case">":o=W(r,t,n);break;case">=":o=tr(r,t,n);break;case"<":o=Y(r,t,n);break;case"<=":o=nr(r,t,n);break;default:throw new TypeError("Invalid operator: "+e)}return o}function ir(r,e){if(r instanceof ir){if(r.loose===e)return r;r=r.value}if(!(this instanceof ir))return new ir(r,e);n("comparator",r,e),this.loose=e,this.parse(r),this.semver===sr?this.value="":this.value=this.operator+this.semver.version,n("comp",this)}e.rcompareIdentifiers=function(r,e){return K(e,r)},e.major=function(r,e){return new H(r,e).major},e.minor=function(r,e){return new H(r,e).minor},e.patch=function(r,e){return new H(r,e).patch},e.compare=Q,e.compareLoose=function(r,e){return Q(r,e,!0)},e.rcompare=function(r,e,t){return Q(e,r,t)},e.sort=function(r,t){return r.sort(function(r,n){return e.compare(r,n,t)})},e.rsort=function(r,t){return r.sort(function(r,n){return e.rcompare(r,n,t)})},e.gt=W,e.lt=Y,e.eq=rr,e.neq=er,e.gte=tr,e.lte=nr,e.cmp=or,e.Comparator=ir;var sr={};function ar(r,e){if(r instanceof ar)return r.loose===e?r:new ar(r.raw,e);if(r instanceof ir)return new ar(r.value,e);if(!(this instanceof ar))return new ar(r,e);if(this.loose=e,this.raw=r,this.set=r.split(/\s*\|\|\s*/).map(function(r){return this.parseRange(r.trim())},this).filter(function(r){return r.length}),!this.set.length)throw new TypeError("Invalid SemVer Range: "+r);this.format()}function ur(r){return!r||"x"===r.toLowerCase()||"*"===r}function cr(r,e,t,n,o,i,s,a,u,c,l,p,h){return((e=ur(t)?"":ur(n)?">="+t+".0.0":ur(o)?">="+t+"."+n+".0":">="+e)+" "+(a=ur(u)?"":ur(c)?"<"+(+u+1)+".0.0":ur(l)?"<"+u+"."+(+c+1)+".0":p?"<="+u+"."+c+"."+l+"-"+p:"<="+a)).trim()}function lr(r,e){for(var t=0;t<r.length;t++)if(!r[t].test(e))return!1;if(e.prerelease.length){for(t=0;t<r.length;t++)if(n(r[t].semver),r[t].semver!==sr&&r[t].semver.prerelease.length>0){var o=r[t].semver;if(o.major===e.major&&o.minor===e.minor&&o.patch===e.patch)return!0}return!1}return!0}function pr(r,e,t){try{e=new ar(e,t)}catch(r){return!1}return e.test(r)}function hr(r,e,t,n){var o,i,s,a,u;switch(r=new H(r,n),e=new ar(e,n),t){case">":o=W,i=nr,s=Y,a=">",u=">=";break;case"<":o=Y,i=tr,s=W,a="<",u="<=";break;default:throw new TypeError('Must provide a hilo val of "<" or ">"')}if(pr(r,e,n))return!1;for(var c=0;c<e.set.length;++c){var l=null,p=null;if(e.set[c].forEach(function(r){r.semver===sr&&(r=new ir(">=0.0.0")),l=l||r,p=p||r,o(r.semver,l.semver,n)?l=r:s(r.semver,p.semver,n)&&(p=r)}),l.operator===a||l.operator===u)return!1;if((!p.operator||p.operator===a)&&i(r,p.semver))return!1;if(p.operator===u&&s(r,p.semver))return!1}return!0}ir.prototype.parse=function(r){var e=this.loose?s[q]:s[X],t=r.match(e);if(!t)throw new TypeError("Invalid comparator: "+r);this.operator=t[1],"="===this.operator&&(this.operator=""),t[2]?this.semver=new H(t[2],this.loose):this.semver=sr},ir.prototype.toString=function(){return this.value},ir.prototype.test=function(r){return n("Comparator.test",r,this.loose),this.semver===sr||("string"==typeof r&&(r=new H(r,this.loose)),or(r,this.operator,this.semver,this.loose))},ir.prototype.intersects=function(r,e){if(!(r instanceof ir))throw new TypeError("a Comparator is required");var t;if(""===this.operator)return t=new ar(r.value,e),pr(this.value,t,e);if(""===r.operator)return t=new ar(this.value,e),pr(r.semver,t,e);var n=!(">="!==this.operator&&">"!==this.operator||">="!==r.operator&&">"!==r.operator),o=!("<="!==this.operator&&"<"!==this.operator||"<="!==r.operator&&"<"!==r.operator),i=this.semver.version===r.semver.version,s=!(">="!==this.operator&&"<="!==this.operator||">="!==r.operator&&"<="!==r.operator),a=or(this.semver,"<",r.semver,e)&&(">="===this.operator||">"===this.operator)&&("<="===r.operator||"<"===r.operator),u=or(this.semver,">",r.semver,e)&&("<="===this.operator||"<"===this.operator)&&(">="===r.operator||">"===r.operator);return n||o||i&&s||a||u},e.Range=ar,ar.prototype.format=function(){return this.range=this.set.map(function(r){return r.join(" ").trim()}).join("||").trim(),this.range},ar.prototype.toString=function(){return this.range},ar.prototype.parseRange=function(r){var e=this.loose;r=r.trim(),n("range",r,e);var t=e?s[G]:s[B];r=r.replace(t,cr),n("hyphen replace",r),r=r.replace(s[z],"$1$2$3"),n("comparator trim",r,s[z]),r=(r=(r=r.replace(s[M],"$1~")).replace(s[C],"$1^")).split(/\s+/).join(" ");var o=e?s[q]:s[X],i=r.split(" ").map(function(r){return function(r,e){return n("comp",r),r=function(r,e){return r.trim().split(/\s+/).map(function(r){return function(r,e){n("caret",r,e);var t=e?s[D]:s[L];return r.replace(t,function(e,t,o,i,s){var a;return n("caret",r,e,t,o,i,s),ur(t)?a="":ur(o)?a=">="+t+".0.0 <"+(+t+1)+".0.0":ur(i)?a="0"===t?">="+t+"."+o+".0 <"+t+"."+(+o+1)+".0":">="+t+"."+o+".0 <"+(+t+1)+".0.0":s?(n("replaceCaret pr",s),"-"!==s.charAt(0)&&(s="-"+s),a="0"===t?"0"===o?">="+t+"."+o+"."+i+s+" <"+t+"."+o+"."+(+i+1):">="+t+"."+o+"."+i+s+" <"+t+"."+(+o+1)+".0":">="+t+"."+o+"."+i+s+" <"+(+t+1)+".0.0"):(n("no pr"),a="0"===t?"0"===o?">="+t+"."+o+"."+i+" <"+t+"."+o+"."+(+i+1):">="+t+"."+o+"."+i+" <"+t+"."+(+o+1)+".0":">="+t+"."+o+"."+i+" <"+(+t+1)+".0.0"),n("caret return",a),a})}(r,e)}).join(" ")}(r,e),n("caret",r),r=function(r,e){return r.trim().split(/\s+/).map(function(r){return function(r,e){var t=e?s[N]:s[V];return r.replace(t,function(e,t,o,i,s){var a;return n("tilde",r,e,t,o,i,s),ur(t)?a="":ur(o)?a=">="+t+".0.0 <"+(+t+1)+".0.0":ur(i)?a=">="+t+"."+o+".0 <"+t+"."+(+o+1)+".0":s?(n("replaceTilde pr",s),"-"!==s.charAt(0)&&(s="-"+s),a=">="+t+"."+o+"."+i+s+" <"+t+"."+(+o+1)+".0"):a=">="+t+"."+o+"."+i+" <"+t+"."+(+o+1)+".0",n("tilde return",a),a})}(r,e)}).join(" ")}(r,e),n("tildes",r),r=function(r,e){return n("replaceXRanges",r,e),r.split(/\s+/).map(function(r){return function(r,e){r=r.trim();var t=e?s[I]:s[R];return r.replace(t,function(e,t,o,i,s,a){n("xRange",r,e,t,o,i,s,a);var u=ur(o),c=u||ur(i),l=c||ur(s),p=l;return"="===t&&p&&(t=""),u?e=">"===t||"<"===t?"<0.0.0":"*":t&&p?(c&&(i=0),l&&(s=0),">"===t?(t=">=",c?(o=+o+1,i=0,s=0):l&&(i=+i+1,s=0)):"<="===t&&(t="<",c?o=+o+1:i=+i+1),e=t+o+"."+i+"."+s):c?e=">="+o+".0.0 <"+(+o+1)+".0.0":l&&(e=">="+o+"."+i+".0 <"+o+"."+(+i+1)+".0"),n("xRange return",e),e})}(r,e)}).join(" ")}(r,e),n("xrange",r),r=function(r,e){return n("replaceStars",r,e),r.trim().replace(s[Z],"")}(r,e),n("stars",r),r}(r,e)}).join(" ").split(/\s+/);return this.loose&&(i=i.filter(function(r){return!!r.match(o)})),i=i.map(function(r){return new ir(r,e)})},ar.prototype.intersects=function(r,e){if(!(r instanceof ar))throw new TypeError("a Range is required");return this.set.some(function(t){return t.every(function(t){return r.set.some(function(r){return r.every(function(r){return t.intersects(r,e)})})})})},e.toComparators=function(r,e){return new ar(r,e).set.map(function(r){return r.map(function(r){return r.value}).join(" ").trim().split(" ")})},ar.prototype.test=function(r){if(!r)return!1;"string"==typeof r&&(r=new H(r,this.loose));for(var e=0;e<this.set.length;e++)if(lr(this.set[e],r))return!0;return!1},e.satisfies=pr,e.maxSatisfying=function(r,e,t){var n=null,o=null;try{var i=new ar(e,t)}catch(r){return null}return r.forEach(function(r){i.test(r)&&(n&&-1!==o.compare(r)||(o=new H(n=r,t)))}),n},e.minSatisfying=function(r,e,t){var n=null,o=null;try{var i=new ar(e,t)}catch(r){return null}return r.forEach(function(r){i.test(r)&&(n&&1!==o.compare(r)||(o=new H(n=r,t)))}),n},e.validRange=function(r,e){try{return new ar(r,e).range||"*"}catch(r){return null}},e.ltr=function(r,e,t){return hr(r,e,"<",t)},e.gtr=function(r,e,t){return hr(r,e,">",t)},e.outside=hr,e.prerelease=function(r,e){var t=F(r,e);return t&&t.prerelease.length?t.prerelease:null},e.intersects=function(r,e,t){return r=new ar(r,t),e=new ar(e,t),r.intersects(e)},e.coerce=function(r){if(r instanceof H)return r;if("string"!=typeof r)return null;var e=r.match(s[O]);return null==e?null:F((e[1]||"0")+"."+(e[2]||"0")+"."+(e[3]||"0"))}}).call(this,t(2))},function(r,e){var t,n,o=r.exports={};function i(){throw new Error("setTimeout has not been defined")}function s(){throw new Error("clearTimeout has not been defined")}function a(r){if(t===setTimeout)return setTimeout(r,0);if((t===i||!t)&&setTimeout)return t=setTimeout,setTimeout(r,0);try{return t(r,0)}catch(e){try{return t.call(null,r,0)}catch(e){return t.call(this,r,0)}}}!function(){try{t="function"==typeof setTimeout?setTimeout:i}catch(r){t=i}try{n="function"==typeof clearTimeout?clearTimeout:s}catch(r){n=s}}();var u,c=[],l=!1,p=-1;function h(){l&&u&&(l=!1,u.length?c=u.concat(c):p=-1,c.length&&f())}function f(){if(!l){var r=a(h);l=!0;for(var e=c.length;e;){for(u=c,c=[];++p<e;)u&&u[p].run();p=-1,e=c.length}u=null,l=!1,function(r){if(n===clearTimeout)return clearTimeout(r);if((n===s||!n)&&clearTimeout)return n=clearTimeout,clearTimeout(r);try{n(r)}catch(e){try{return n.call(null,r)}catch(e){return n.call(this,r)}}}(r)}}function v(r,e){this.fun=r,this.array=e}function m(){}o.nextTick=function(r){var e=new Array(arguments.length-1);if(arguments.length>1)for(var t=1;t<arguments.length;t++)e[t-1]=arguments[t];c.push(new v(r,e)),1!==c.length||l||a(f)},v.prototype.run=function(){this.fun.apply(null,this.array)},o.title="browser",o.browser=!0,o.env={},o.argv=[],o.version="",o.versions={},o.on=m,o.addListener=m,o.once=m,o.off=m,o.removeListener=m,o.removeAllListeners=m,o.emit=m,o.prependListener=m,o.prependOnceListener=m,o.listeners=function(r){return[]},o.binding=function(r){throw new Error("process.binding is not supported")},o.cwd=function(){return"/"},o.chdir=function(r){throw new Error("process.chdir is not supported")},o.umask=function(){return 0}}]);
+!function(r){var e={};function t(n){if(e[n])return e[n].exports;var o=e[n]={i:n,l:!1,exports:{}};return r[n].call(o.exports,o,o.exports,t),o.l=!0,o.exports}t.m=r,t.c=e,t.d=function(r,e,n){t.o(r,e)||Object.defineProperty(r,e,{enumerable:!0,get:n})},t.r=function(r){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(r,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(r,"__esModule",{value:!0})},t.t=function(r,e){if(1&e&&(r=t(r)),8&e)return r;if(4&e&&"object"==typeof r&&r&&r.__esModule)return r;var n=Object.create(null);if(t.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:r}),2&e&&"string"!=typeof r)for(var o in r)t.d(n,o,function(e){return r[e]}.bind(null,o));return n},t.n=function(r){var e=r&&r.__esModule?function(){return r.default}:function(){return r};return t.d(e,"a",e),e},t.o=function(r,e){return Object.prototype.hasOwnProperty.call(r,e)},t.p="",t(t.s=0)}([function(r,e,t){const n=t(1);function o(r){var e=r.version;console.debug("Running version: "+e.slug);var t={project__slug:r.project.slug};$.ajax({url:r.meta.api_url+"version/",data:t,success:function(t){highest_version=function(r){var e;return $.each(r,function(r,t){n.valid(t.slug)&&(e?n.valid(t.slug)&&n.valid(e.slug)&&n.gt(t.slug,e.slug)&&(e=t):e=t)}),e}(t.results),n.valid(e.slug)&&n.valid(highest_version.slug)&&n.lt(e.slug,highest_version.slug)&&(console.debug("Highest version: "+highest_version.slug),function(r,e,t){var n=window.location.pathname.replace(r.slug,e.slug),o=$(t.banner.html);o.find("a").attr("href",n).text(e.slug),$(t.banner.body_default_selector).prepend(o)}(e,highest_version,r))},error:function(){console.error("Error loading Read the Docs active versions.")}})}$(document).ready(function(){$.ajax({url:"_static/data/versionwarning-data.json",success:function(r){document.getElementById(r.banner.id_div)?console.debug("There is already a banner added. No checking versions."):o(r)},error:function(){console.error("Error loading versionwarning-data.json")}})})},function(r,e,t){(function(t){var n;e=r.exports=H,n="object"==typeof t&&t.env&&t.env.NODE_DEBUG&&/\bsemver\b/i.test(t.env.NODE_DEBUG)?function(){var r=Array.prototype.slice.call(arguments,0);r.unshift("SEMVER"),console.log.apply(console,r)}:function(){},e.SEMVER_SPEC_VERSION="2.0.0";var o=256,i=Number.MAX_SAFE_INTEGER||9007199254740991,s=e.re=[],a=e.src=[],u=0,c=u++;a[c]="0|[1-9]\\d*";var l=u++;a[l]="[0-9]+";var p=u++;a[p]="\\d*[a-zA-Z-][a-zA-Z0-9-]*";var h=u++;a[h]="("+a[c]+")\\.("+a[c]+")\\.("+a[c]+")";var f=u++;a[f]="("+a[l]+")\\.("+a[l]+")\\.("+a[l]+")";var v=u++;a[v]="(?:"+a[c]+"|"+a[p]+")";var m=u++;a[m]="(?:"+a[l]+"|"+a[p]+")";var g=u++;a[g]="(?:-("+a[v]+"(?:\\."+a[v]+")*))";var d=u++;a[d]="(?:-?("+a[m]+"(?:\\."+a[m]+")*))";var w=u++;a[w]="[0-9A-Za-z-]+";var y=u++;a[y]="(?:\\+("+a[w]+"(?:\\."+a[w]+")*))";var b=u++,j="v?"+a[h]+a[g]+"?"+a[y]+"?";a[b]="^"+j+"$";var E="[v=\\s]*"+a[f]+a[d]+"?"+a[y]+"?",T=u++;a[T]="^"+E+"$";var $=u++;a[$]="((?:<|>)?=?)";var _=u++;a[_]=a[l]+"|x|X|\\*";var x=u++;a[x]=a[c]+"|x|X|\\*";var k=u++;a[k]="[v=\\s]*("+a[x]+")(?:\\.("+a[x]+")(?:\\.("+a[x]+")(?:"+a[g]+")?"+a[y]+"?)?)?";var S=u++;a[S]="[v=\\s]*("+a[_]+")(?:\\.("+a[_]+")(?:\\.("+a[_]+")(?:"+a[d]+")?"+a[y]+"?)?)?";var R=u++;a[R]="^"+a[$]+"\\s*"+a[k]+"$";var I=u++;a[I]="^"+a[$]+"\\s*"+a[S]+"$";var O=u++;a[O]="(?:^|[^\\d])(\\d{1,16})(?:\\.(\\d{1,16}))?(?:\\.(\\d{1,16}))?(?:$|[^\\d])";var A=u++;a[A]="(?:~>?)";var M=u++;a[M]="(\\s*)"+a[A]+"\\s+",s[M]=new RegExp(a[M],"g");var V=u++;a[V]="^"+a[A]+a[k]+"$";var N=u++;a[N]="^"+a[A]+a[S]+"$";var P=u++;a[P]="(?:\\^)";var C=u++;a[C]="(\\s*)"+a[P]+"\\s+",s[C]=new RegExp(a[C],"g");var L=u++;a[L]="^"+a[P]+a[k]+"$";var D=u++;a[D]="^"+a[P]+a[S]+"$";var q=u++;a[q]="^"+a[$]+"\\s*("+E+")$|^$";var X=u++;a[X]="^"+a[$]+"\\s*("+j+")$|^$";var z=u++;a[z]="(\\s*)"+a[$]+"\\s*("+E+"|"+a[k]+")",s[z]=new RegExp(a[z],"g");var B=u++;a[B]="^\\s*("+a[k]+")\\s+-\\s+("+a[k]+")\\s*$";var G=u++;a[G]="^\\s*("+a[S]+")\\s+-\\s+("+a[S]+")\\s*$";var Z=u++;a[Z]="(<|>)?=?\\s*\\*";for(var U=0;U<u;U++)n(U,a[U]),s[U]||(s[U]=new RegExp(a[U]));function F(r,e){if(r instanceof H)return r;if("string"!=typeof r)return null;if(r.length>o)return null;if(!(e?s[T]:s[b]).test(r))return null;try{return new H(r,e)}catch(r){return null}}function H(r,e){if(r instanceof H){if(r.loose===e)return r;r=r.version}else if("string"!=typeof r)throw new TypeError("Invalid Version: "+r);if(r.length>o)throw new TypeError("version is longer than "+o+" characters");if(!(this instanceof H))return new H(r,e);n("SemVer",r,e),this.loose=e;var t=r.trim().match(e?s[T]:s[b]);if(!t)throw new TypeError("Invalid Version: "+r);if(this.raw=r,this.major=+t[1],this.minor=+t[2],this.patch=+t[3],this.major>i||this.major<0)throw new TypeError("Invalid major version");if(this.minor>i||this.minor<0)throw new TypeError("Invalid minor version");if(this.patch>i||this.patch<0)throw new TypeError("Invalid patch version");t[4]?this.prerelease=t[4].split(".").map(function(r){if(/^[0-9]+$/.test(r)){var e=+r;if(e>=0&&e<i)return e}return r}):this.prerelease=[],this.build=t[5]?t[5].split("."):[],this.format()}e.parse=F,e.valid=function(r,e){var t=F(r,e);return t?t.version:null},e.clean=function(r,e){var t=F(r.trim().replace(/^[=v]+/,""),e);return t?t.version:null},e.SemVer=H,H.prototype.format=function(){return this.version=this.major+"."+this.minor+"."+this.patch,this.prerelease.length&&(this.version+="-"+this.prerelease.join(".")),this.version},H.prototype.toString=function(){return this.version},H.prototype.compare=function(r){return n("SemVer.compare",this.version,this.loose,r),r instanceof H||(r=new H(r,this.loose)),this.compareMain(r)||this.comparePre(r)},H.prototype.compareMain=function(r){return r instanceof H||(r=new H(r,this.loose)),K(this.major,r.major)||K(this.minor,r.minor)||K(this.patch,r.patch)},H.prototype.comparePre=function(r){if(r instanceof H||(r=new H(r,this.loose)),this.prerelease.length&&!r.prerelease.length)return-1;if(!this.prerelease.length&&r.prerelease.length)return 1;if(!this.prerelease.length&&!r.prerelease.length)return 0;var e=0;do{var t=this.prerelease[e],o=r.prerelease[e];if(n("prerelease compare",e,t,o),void 0===t&&void 0===o)return 0;if(void 0===o)return 1;if(void 0===t)return-1;if(t!==o)return K(t,o)}while(++e)},H.prototype.inc=function(r,e){switch(r){case"premajor":this.prerelease.length=0,this.patch=0,this.minor=0,this.major++,this.inc("pre",e);break;case"preminor":this.prerelease.length=0,this.patch=0,this.minor++,this.inc("pre",e);break;case"prepatch":this.prerelease.length=0,this.inc("patch",e),this.inc("pre",e);break;case"prerelease":0===this.prerelease.length&&this.inc("patch",e),this.inc("pre",e);break;case"major":0===this.minor&&0===this.patch&&0!==this.prerelease.length||this.major++,this.minor=0,this.patch=0,this.prerelease=[];break;case"minor":0===this.patch&&0!==this.prerelease.length||this.minor++,this.patch=0,this.prerelease=[];break;case"patch":0===this.prerelease.length&&this.patch++,this.prerelease=[];break;case"pre":if(0===this.prerelease.length)this.prerelease=[0];else{for(var t=this.prerelease.length;--t>=0;)"number"==typeof this.prerelease[t]&&(this.prerelease[t]++,t=-2);-1===t&&this.prerelease.push(0)}e&&(this.prerelease[0]===e?isNaN(this.prerelease[1])&&(this.prerelease=[e,0]):this.prerelease=[e,0]);break;default:throw new Error("invalid increment argument: "+r)}return this.format(),this.raw=this.version,this},e.inc=function(r,e,t,n){"string"==typeof t&&(n=t,t=void 0);try{return new H(r,t).inc(e,n).version}catch(r){return null}},e.diff=function(r,e){if(rr(r,e))return null;var t=F(r),n=F(e);if(t.prerelease.length||n.prerelease.length){for(var o in t)if(("major"===o||"minor"===o||"patch"===o)&&t[o]!==n[o])return"pre"+o;return"prerelease"}for(var o in t)if(("major"===o||"minor"===o||"patch"===o)&&t[o]!==n[o])return o},e.compareIdentifiers=K;var J=/^[0-9]+$/;function K(r,e){var t=J.test(r),n=J.test(e);return t&&n&&(r=+r,e=+e),t&&!n?-1:n&&!t?1:r<e?-1:r>e?1:0}function Q(r,e,t){return new H(r,t).compare(new H(e,t))}function W(r,e,t){return Q(r,e,t)>0}function Y(r,e,t){return Q(r,e,t)<0}function rr(r,e,t){return 0===Q(r,e,t)}function er(r,e,t){return 0!==Q(r,e,t)}function tr(r,e,t){return Q(r,e,t)>=0}function nr(r,e,t){return Q(r,e,t)<=0}function or(r,e,t,n){var o;switch(e){case"===":"object"==typeof r&&(r=r.version),"object"==typeof t&&(t=t.version),o=r===t;break;case"!==":"object"==typeof r&&(r=r.version),"object"==typeof t&&(t=t.version),o=r!==t;break;case"":case"=":case"==":o=rr(r,t,n);break;case"!=":o=er(r,t,n);break;case">":o=W(r,t,n);break;case">=":o=tr(r,t,n);break;case"<":o=Y(r,t,n);break;case"<=":o=nr(r,t,n);break;default:throw new TypeError("Invalid operator: "+e)}return o}function ir(r,e){if(r instanceof ir){if(r.loose===e)return r;r=r.value}if(!(this instanceof ir))return new ir(r,e);n("comparator",r,e),this.loose=e,this.parse(r),this.semver===sr?this.value="":this.value=this.operator+this.semver.version,n("comp",this)}e.rcompareIdentifiers=function(r,e){return K(e,r)},e.major=function(r,e){return new H(r,e).major},e.minor=function(r,e){return new H(r,e).minor},e.patch=function(r,e){return new H(r,e).patch},e.compare=Q,e.compareLoose=function(r,e){return Q(r,e,!0)},e.rcompare=function(r,e,t){return Q(e,r,t)},e.sort=function(r,t){return r.sort(function(r,n){return e.compare(r,n,t)})},e.rsort=function(r,t){return r.sort(function(r,n){return e.rcompare(r,n,t)})},e.gt=W,e.lt=Y,e.eq=rr,e.neq=er,e.gte=tr,e.lte=nr,e.cmp=or,e.Comparator=ir;var sr={};function ar(r,e){if(r instanceof ar)return r.loose===e?r:new ar(r.raw,e);if(r instanceof ir)return new ar(r.value,e);if(!(this instanceof ar))return new ar(r,e);if(this.loose=e,this.raw=r,this.set=r.split(/\s*\|\|\s*/).map(function(r){return this.parseRange(r.trim())},this).filter(function(r){return r.length}),!this.set.length)throw new TypeError("Invalid SemVer Range: "+r);this.format()}function ur(r){return!r||"x"===r.toLowerCase()||"*"===r}function cr(r,e,t,n,o,i,s,a,u,c,l,p,h){return((e=ur(t)?"":ur(n)?">="+t+".0.0":ur(o)?">="+t+"."+n+".0":">="+e)+" "+(a=ur(u)?"":ur(c)?"<"+(+u+1)+".0.0":ur(l)?"<"+u+"."+(+c+1)+".0":p?"<="+u+"."+c+"."+l+"-"+p:"<="+a)).trim()}function lr(r,e){for(var t=0;t<r.length;t++)if(!r[t].test(e))return!1;if(e.prerelease.length){for(t=0;t<r.length;t++)if(n(r[t].semver),r[t].semver!==sr&&r[t].semver.prerelease.length>0){var o=r[t].semver;if(o.major===e.major&&o.minor===e.minor&&o.patch===e.patch)return!0}return!1}return!0}function pr(r,e,t){try{e=new ar(e,t)}catch(r){return!1}return e.test(r)}function hr(r,e,t,n){var o,i,s,a,u;switch(r=new H(r,n),e=new ar(e,n),t){case">":o=W,i=nr,s=Y,a=">",u=">=";break;case"<":o=Y,i=tr,s=W,a="<",u="<=";break;default:throw new TypeError('Must provide a hilo val of "<" or ">"')}if(pr(r,e,n))return!1;for(var c=0;c<e.set.length;++c){var l=null,p=null;if(e.set[c].forEach(function(r){r.semver===sr&&(r=new ir(">=0.0.0")),l=l||r,p=p||r,o(r.semver,l.semver,n)?l=r:s(r.semver,p.semver,n)&&(p=r)}),l.operator===a||l.operator===u)return!1;if((!p.operator||p.operator===a)&&i(r,p.semver))return!1;if(p.operator===u&&s(r,p.semver))return!1}return!0}ir.prototype.parse=function(r){var e=this.loose?s[q]:s[X],t=r.match(e);if(!t)throw new TypeError("Invalid comparator: "+r);this.operator=t[1],"="===this.operator&&(this.operator=""),t[2]?this.semver=new H(t[2],this.loose):this.semver=sr},ir.prototype.toString=function(){return this.value},ir.prototype.test=function(r){return n("Comparator.test",r,this.loose),this.semver===sr||("string"==typeof r&&(r=new H(r,this.loose)),or(r,this.operator,this.semver,this.loose))},ir.prototype.intersects=function(r,e){if(!(r instanceof ir))throw new TypeError("a Comparator is required");var t;if(""===this.operator)return t=new ar(r.value,e),pr(this.value,t,e);if(""===r.operator)return t=new ar(this.value,e),pr(r.semver,t,e);var n=!(">="!==this.operator&&">"!==this.operator||">="!==r.operator&&">"!==r.operator),o=!("<="!==this.operator&&"<"!==this.operator||"<="!==r.operator&&"<"!==r.operator),i=this.semver.version===r.semver.version,s=!(">="!==this.operator&&"<="!==this.operator||">="!==r.operator&&"<="!==r.operator),a=or(this.semver,"<",r.semver,e)&&(">="===this.operator||">"===this.operator)&&("<="===r.operator||"<"===r.operator),u=or(this.semver,">",r.semver,e)&&("<="===this.operator||"<"===this.operator)&&(">="===r.operator||">"===r.operator);return n||o||i&&s||a||u},e.Range=ar,ar.prototype.format=function(){return this.range=this.set.map(function(r){return r.join(" ").trim()}).join("||").trim(),this.range},ar.prototype.toString=function(){return this.range},ar.prototype.parseRange=function(r){var e=this.loose;r=r.trim(),n("range",r,e);var t=e?s[G]:s[B];r=r.replace(t,cr),n("hyphen replace",r),r=r.replace(s[z],"$1$2$3"),n("comparator trim",r,s[z]),r=(r=(r=r.replace(s[M],"$1~")).replace(s[C],"$1^")).split(/\s+/).join(" ");var o=e?s[q]:s[X],i=r.split(" ").map(function(r){return function(r,e){return n("comp",r),r=function(r,e){return r.trim().split(/\s+/).map(function(r){return function(r,e){n("caret",r,e);var t=e?s[D]:s[L];return r.replace(t,function(e,t,o,i,s){var a;return n("caret",r,e,t,o,i,s),ur(t)?a="":ur(o)?a=">="+t+".0.0 <"+(+t+1)+".0.0":ur(i)?a="0"===t?">="+t+"."+o+".0 <"+t+"."+(+o+1)+".0":">="+t+"."+o+".0 <"+(+t+1)+".0.0":s?(n("replaceCaret pr",s),"-"!==s.charAt(0)&&(s="-"+s),a="0"===t?"0"===o?">="+t+"."+o+"."+i+s+" <"+t+"."+o+"."+(+i+1):">="+t+"."+o+"."+i+s+" <"+t+"."+(+o+1)+".0":">="+t+"."+o+"."+i+s+" <"+(+t+1)+".0.0"):(n("no pr"),a="0"===t?"0"===o?">="+t+"."+o+"."+i+" <"+t+"."+o+"."+(+i+1):">="+t+"."+o+"."+i+" <"+t+"."+(+o+1)+".0":">="+t+"."+o+"."+i+" <"+(+t+1)+".0.0"),n("caret return",a),a})}(r,e)}).join(" ")}(r,e),n("caret",r),r=function(r,e){return r.trim().split(/\s+/).map(function(r){return function(r,e){var t=e?s[N]:s[V];return r.replace(t,function(e,t,o,i,s){var a;return n("tilde",r,e,t,o,i,s),ur(t)?a="":ur(o)?a=">="+t+".0.0 <"+(+t+1)+".0.0":ur(i)?a=">="+t+"."+o+".0 <"+t+"."+(+o+1)+".0":s?(n("replaceTilde pr",s),"-"!==s.charAt(0)&&(s="-"+s),a=">="+t+"."+o+"."+i+s+" <"+t+"."+(+o+1)+".0"):a=">="+t+"."+o+"."+i+" <"+t+"."+(+o+1)+".0",n("tilde return",a),a})}(r,e)}).join(" ")}(r,e),n("tildes",r),r=function(r,e){return n("replaceXRanges",r,e),r.split(/\s+/).map(function(r){return function(r,e){r=r.trim();var t=e?s[I]:s[R];return r.replace(t,function(e,t,o,i,s,a){n("xRange",r,e,t,o,i,s,a);var u=ur(o),c=u||ur(i),l=c||ur(s),p=l;return"="===t&&p&&(t=""),u?e=">"===t||"<"===t?"<0.0.0":"*":t&&p?(c&&(i=0),l&&(s=0),">"===t?(t=">=",c?(o=+o+1,i=0,s=0):l&&(i=+i+1,s=0)):"<="===t&&(t="<",c?o=+o+1:i=+i+1),e=t+o+"."+i+"."+s):c?e=">="+o+".0.0 <"+(+o+1)+".0.0":l&&(e=">="+o+"."+i+".0 <"+o+"."+(+i+1)+".0"),n("xRange return",e),e})}(r,e)}).join(" ")}(r,e),n("xrange",r),r=function(r,e){return n("replaceStars",r,e),r.trim().replace(s[Z],"")}(r,e),n("stars",r),r}(r,e)}).join(" ").split(/\s+/);return this.loose&&(i=i.filter(function(r){return!!r.match(o)})),i=i.map(function(r){return new ir(r,e)})},ar.prototype.intersects=function(r,e){if(!(r instanceof ar))throw new TypeError("a Range is required");return this.set.some(function(t){return t.every(function(t){return r.set.some(function(r){return r.every(function(r){return t.intersects(r,e)})})})})},e.toComparators=function(r,e){return new ar(r,e).set.map(function(r){return r.map(function(r){return r.value}).join(" ").trim().split(" ")})},ar.prototype.test=function(r){if(!r)return!1;"string"==typeof r&&(r=new H(r,this.loose));for(var e=0;e<this.set.length;e++)if(lr(this.set[e],r))return!0;return!1},e.satisfies=pr,e.maxSatisfying=function(r,e,t){var n=null,o=null;try{var i=new ar(e,t)}catch(r){return null}return r.forEach(function(r){i.test(r)&&(n&&-1!==o.compare(r)||(o=new H(n=r,t)))}),n},e.minSatisfying=function(r,e,t){var n=null,o=null;try{var i=new ar(e,t)}catch(r){return null}return r.forEach(function(r){i.test(r)&&(n&&1!==o.compare(r)||(o=new H(n=r,t)))}),n},e.validRange=function(r,e){try{return new ar(r,e).range||"*"}catch(r){return null}},e.ltr=function(r,e,t){return hr(r,e,"<",t)},e.gtr=function(r,e,t){return hr(r,e,">",t)},e.outside=hr,e.prerelease=function(r,e){var t=F(r,e);return t&&t.prerelease.length?t.prerelease:null},e.intersects=function(r,e,t){return r=new ar(r,t),e=new ar(e,t),r.intersects(e)},e.coerce=function(r){if(r instanceof H)return r;if("string"!=typeof r)return null;var e=r.match(s[O]);return null==e?null:F((e[1]||"0")+"."+(e[2]||"0")+"."+(e[3]||"0"))}}).call(this,t(2))},function(r,e){var t,n,o=r.exports={};function i(){throw new Error("setTimeout has not been defined")}function s(){throw new Error("clearTimeout has not been defined")}function a(r){if(t===setTimeout)return setTimeout(r,0);if((t===i||!t)&&setTimeout)return t=setTimeout,setTimeout(r,0);try{return t(r,0)}catch(e){try{return t.call(null,r,0)}catch(e){return t.call(this,r,0)}}}!function(){try{t="function"==typeof setTimeout?setTimeout:i}catch(r){t=i}try{n="function"==typeof clearTimeout?clearTimeout:s}catch(r){n=s}}();var u,c=[],l=!1,p=-1;function h(){l&&u&&(l=!1,u.length?c=u.concat(c):p=-1,c.length&&f())}function f(){if(!l){var r=a(h);l=!0;for(var e=c.length;e;){for(u=c,c=[];++p<e;)u&&u[p].run();p=-1,e=c.length}u=null,l=!1,function(r){if(n===clearTimeout)return clearTimeout(r);if((n===s||!n)&&clearTimeout)return n=clearTimeout,clearTimeout(r);try{n(r)}catch(e){try{return n.call(null,r)}catch(e){return n.call(this,r)}}}(r)}}function v(r,e){this.fun=r,this.array=e}function m(){}o.nextTick=function(r){var e=new Array(arguments.length-1);if(arguments.length>1)for(var t=1;t<arguments.length;t++)e[t-1]=arguments[t];c.push(new v(r,e)),1!==c.length||l||a(f)},v.prototype.run=function(){this.fun.apply(null,this.array)},o.title="browser",o.browser=!0,o.env={},o.argv=[],o.version="",o.versions={},o.on=m,o.addListener=m,o.once=m,o.off=m,o.removeListener=m,o.removeAllListeners=m,o.emit=m,o.prependListener=m,o.prependOnceListener=m,o.listeners=function(r){return[]},o.binding=function(r){throw new Error("process.binding is not supported")},o.cwd=function(){return"/"},o.chdir=function(r){throw new Error("process.chdir is not supported")},o.umask=function(){return 0}}]);
