+++ versionwarning/signals.py
@@ -1,24 +1,10 @@
 # -*- coding: utf-8 -*-
 
-from docutils import nodes
 import json
 import os
 
-from .banner import VersionWarningBanner
 
 STATIC_PATH = os.path.join(os.path.dirname(__file__), '_static')
-
-
-def process_version_warning_banner(app, doctree, fromdocname):
-    """
-    Insert an ``admonition`` with a custom/general message.
-    """
-
-    banner = VersionWarningBanner(app, doctree)
-    for document in doctree.traverse(nodes.document):
-        banner = banner.get_banner_node()
-        if isinstance(banner, nodes.Admonition):
-            document.insert(0, banner)
 
 
 def generate_versionwarning_data_json(app, **kwargs):
@@ -40,23 +26,31 @@ def generate_versionwarning_data_json(app, **kwargs):
     if config is None:
         config = app.config
 
+    banner_html = config.versionwarning_banner_html.format(
+        id_div=config.versionwarning_banner_id_div,
+        banner_title=config.versionwarning_banner_title,
+        message=config.versionwarning_default_message.format(
+            **{config.versionwarning_message_placeholder: '<a href="#"></a>'},
+        ),
+        admonition_type=config.versionwarning_admonition_type,
+    )
+
     data = json.dumps({
         'meta': {
             'api_url': config.versionwarning_api_url,
         },
         'banner': {
-            'html': config.versionwarning_banner_html,
+            'html': banner_html,
             'id_div': config.versionwarning_banner_id_div,
             'body_default_selector': config.versionwarning_body_default_selector,
-            'body_extra_selector': config.versionwarning_body_extra_selector,
         },
         'project': {
             'slug': config.versionwarning_project_slug,
         },
         'version': {
-            'slug': config.version,
+            'slug': config.versionwarning_project_version,
         },
-    })
+    }, indent=4)
 
     data_path = os.path.join(STATIC_PATH, 'data')
     if not os.path.exists(data_path):
